refactor: 결과가 정수가 아닐 경우를 처리
int -> double로 변경

+++ src/main/java/unit/domain/Calculator.java
@@ -7,7 +7,7 @@ public class Calculator {
 		this.expression = expression;
 	}
 
-	public int getResult() {
+	public double getResult() {
 		return expression.calculate();
 	}
 }
