+++ src/main/java/unit/view/OutputView.java
@@ -1,7 +1,7 @@
 package unit.view;
 
 public class OutputView {
-	public static void printResult(final int result) {
-		System.out.println(String.format("계산 결과 : %d", result));
+	public static void printResult(final double result) {
+		System.out.println(String.format("계산 결과 : %.3f", result));
 	}
 }
